+++ library/data-store/src/main/java/io/mehow/laboratory/datastore/DataStoreFeatureStorage.kt
@@ -2,6 +2,7 @@ package io.mehow.laboratory.datastore
 
 import android.content.Context
 import androidx.datastore.core.DataStore
+import androidx.datastore.core.DataStoreFactory
 import io.mehow.laboratory.Feature
 import io.mehow.laboratory.FeatureStorage
 import kotlinx.coroutines.flow.first
@@ -51,12 +52,12 @@ public fun FeatureStorage.Companion.dataStore(dataStore: DataStore<FeatureFlags>
  * Creates a [FeatureStorage] that is backed by [DataStore] with a file taken from [fileProvider].
  */
 public fun FeatureStorage.Companion.dataStore(fileProvider: () -> File): FeatureStorage {
-  return dataStoreBuilder(fileProvider).build()
+  return dataStore(DataStoreFactory.create(FeatureFlagsSerializer, produceFile = fileProvider))
 }
 
 /**
  * Creates a [FeatureStorage] that is backed by [DataStore] with a file in an apps default directory.
  */
 public fun FeatureStorage.Companion.dataStore(context: Context, fileName: String): FeatureStorage {
-  return dataStoreBuilder(context, fileName).build()
+  return dataStore { File(context.filesDir, "datastore/$fileName") }
 }